Fixing the defect of a problem where any user can edit and delete any…

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -36,9 +36,9 @@ public class ImageController {
     }
 
     //This method is called when the details of the specific image with corresponding title are to be displayed
-    //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
-    //First receive the dynamic parameter in the incoming request URL in a string variable 'title' and also the Model type object
-    //Call the getImageByTitle() method in the business logic to fetch all the details of that image
+    //The logic is to get the image from the database with corresponding title. After getting the image from the database the details are shown
+    //First receive the dynamic parameter in the incoming request URL in a Integer variable image id and also the Model type object
+    //Call the getImageById() method in the business logic to fetch all the details of that image
     //Add the image in the Model type object with 'image' as the key
     //Return 'images/image.html' file
 
@@ -91,14 +91,27 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    
+    /*To handle only image owner can edit image, first of all it is getting image object and getting the owner's id (user id)
+      Next we are getting the current logged in user by HttpSession
+      Next we compare both IDs, if both are same we will navigate to edit page otherwise we will show the error and stay on the same page*/ 
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        String editError = "Only the owner of the image can edit the image";
+        Integer imgOwner = image.getUser().getId();
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+        if (imgOwner == loggedInUser.getId()) {
+          String tags = convertTagsToString(image.getTags());
+          model.addAttribute("image", image);
+          model.addAttribute("tags", tags);
+          return "images/edit";
+        } else {
+          model.addAttribute("editError", editError);
+          model.addAttribute("image", image);
+          model.addAttribute("tags", image.getTags());
+          return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -110,7 +123,7 @@ public class ImageController {
     //Call the updateImage() method in the business logic to update the image
     //Direct to the same page showing the details of that particular updated image
 
-    //The method also receives tags parameter which is a string of all the tags separated by a comma using the annotation @RequestParam
+    //The method receives tags parameter which is a string of all the tags separated by a comma using the annotation @RequestParam
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
     public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
@@ -139,10 +152,27 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    
+    /*To handle only image owner can delete that image, first of all it is getting image object and getting the owner's id (user id)
+      Next we are getting the current logged in user by HttpSession
+      Next we compare both IDs, if both are same we will redirect the request to images page otherwise we will show the delete error and stay on the same page*/ 
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        Integer imgOwner = image.getUser().getId();
+        String deleteError = "Only the owner of the image can delete the image";
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+        image.setUser(loggedInUser);
+        List<Tag> tags = image.getTags();
+        if (imgOwner == loggedInUser.getId()) {
+          imageService.deleteImage(imageId);
+          return "redirect:/images";
+        } else {
+          model.addAttribute("deleteError", deleteError);
+          model.addAttribute("image", image);
+          model.addAttribute("tags", tags);
+          return "images/image";
+        }
     }
